+++ internal/common/apps.go
@@ -14,9 +14,9 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/giantswarm/clustertest/v2/pkg/failurehandler"
-	"github.com/giantswarm/clustertest/v2/pkg/logger"
-	"github.com/giantswarm/clustertest/v2/pkg/wait"
+	"github.com/giantswarm/clustertest/v3/pkg/failurehandler"
+	"github.com/giantswarm/clustertest/v3/pkg/logger"
+	"github.com/giantswarm/clustertest/v3/pkg/wait"
 
 	"github.com/giantswarm/cluster-test-suites/v2/internal/helper"
 	"github.com/giantswarm/cluster-test-suites/v2/internal/state"
@@ -59,26 +59,13 @@ func RunApps() {
 		})
 
 		It("all default apps are deployed without issues", func() {
-			skipDefaultAppsApp, err := state.GetCluster().UsesUnifiedClusterApp()
-			Expect(err).NotTo(HaveOccurred())
-
 			timeout := state.GetTestTimeout(timeout.DeployApps, 15*time.Minute)
 			logger.Log("Waiting for all apps to be deployed. Timeout: %s", timeout.String())
-
-			if skipDefaultAppsApp {
-				logger.Log("Checking default apps deployed from the unified %s app (with default apps), so skipping check of %s App resource as it does not exist.", state.GetCluster().ClusterApp.AppName, state.GetCluster().DefaultAppsApp.AppName)
-			} else {
-				// We need to wait for default-apps to be deployed before we can check all apps.
-				defaultAppsAppName := fmt.Sprintf("%s-%s", state.GetCluster().Name, "default-apps")
-				Eventually(wait.IsAppDeployed(state.GetContext(), state.GetFramework().MC(), defaultAppsAppName, state.GetCluster().Organization.GetNamespace())).
-					WithTimeout(30 * time.Second).
-					WithPolling(50 * time.Millisecond).
-					Should(BeTrue())
-			}
+			logger.Log("Checking default apps deployed from the unified %s app.", state.GetCluster().ClusterApp.AppName)
 
 			// Wait for all default-apps apps to be deployed
 			appList := &v1alpha1.AppList{}
-			err = state.GetFramework().MC().List(state.GetContext(), appList, ctrl.InNamespace(state.GetCluster().Organization.GetNamespace()), getDefaultAppsSelector())
+			err := state.GetFramework().MC().List(state.GetContext(), appList, ctrl.InNamespace(state.GetCluster().Organization.GetNamespace()), getDefaultAppsSelector())
 			Expect(err).NotTo(HaveOccurred())
 
 			appNamespacedNames := []types.NamespacedName{}
@@ -164,22 +151,11 @@ func RunApps() {
 }
 
 func getDefaultAppsSelector() ctrl.MatchingLabels {
-	defaultAppsAppName := fmt.Sprintf("%s-%s", state.GetCluster().Name, "default-apps")
-	skipDefaultAppsApp, err := state.GetCluster().UsesUnifiedClusterApp()
-	Expect(err).NotTo(HaveOccurred())
-
-	var defaultAppsSelectorLabels ctrl.MatchingLabels
-	if skipDefaultAppsApp {
-		defaultAppsSelectorLabels = ctrl.MatchingLabels{
-			"giantswarm.io/cluster":        state.GetCluster().Name,
-			"app.kubernetes.io/managed-by": "Helm",
-		}
-	} else {
-		defaultAppsSelectorLabels = ctrl.MatchingLabels{
-			"giantswarm.io/managed-by": defaultAppsAppName,
-		}
+	// All providers now use unified cluster apps that deploy default apps directly via Helm
+	return ctrl.MatchingLabels{
+		"giantswarm.io/cluster":        state.GetCluster().Name,
+		"app.kubernetes.io/managed-by": "Helm",
 	}
-	return defaultAppsSelectorLabels
 }
 
 func reportOwningTeams() failurehandler.FailureHandler {