+++ providers/capv/on-capz/capv_suite_test.go
@@ -6,7 +6,7 @@ import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 	. "github.com/onsi/gomega"    //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-standup-teardown/v2/pkg/clusterbuilder/providers/capv"
+	"github.com/giantswarm/cluster-standup-teardown/v4/pkg/clusterbuilder/providers/capv"
 
 	"github.com/giantswarm/cluster-test-suites/v2/internal/suite"
 )